Fix floating point comparisons in KEMO animations.
The floating point comparisons in kemo_animation_hb were not safe.
In some cases, due to floating point imprecision, they could result
in a model scale of nearly zero being set on a PC's model in
QuickNormalizeHeight.

The fix is to introduce proper comparisons with an epsilon value to
account for imprecision.