Add ScriptLoader script (dynamically load and run new C# scripts).
The ScriptLoader script can be used to dynamically load and run a
new script assembly, or a modified version of a script assembly.

The newly loaded script does not replace existing instances for
purposes of normal ExecuteScript() or ExecuteScriptEnhanced()
calls.  But it can be used to test the functionality of the script
without requiring a server reboot.  It can also be used to perform
emergency changes to a running server instance by injecting and
running custom script code.

To use the script loader (from NWScript):

//
// Add the full path to the .DLL to load
// This might be best exposed as a chat command that DMs can
// issue.
//

AddScriptParameterString("C:\\ALFA\\TestScript.dll");
ExecuteScriptEnhanced("scriptloader", TargetObject, TRUE);

This will load and call the ScriptMain of TestScript.dll, so that
the functionality can be tested without a server reboot.  The
loaded script object doesn't supersede existing instances, but it
can start DelayCommands or other continuations.

- Enhanced ALFA.ScriptLoader to support loading a new CLR script
  assembly in from disk.

- Enhanced ALFA.ScriptLoader to support making the default
  parameter list for a script object.

- Added ScriptLoader wrapper script around ALFA.ScriptLoader.  This
  script will load the specified CLR script assembly (given by the
  first string argument) and call its entrypoint with default
  parameters.