Maintain Module.Areas properly, add #sa rs $CreatureAI debugging.
$CreatureAI.GetCreatures() => Get a list of creatures known to AI.
$CreatureAI.GetAreas() => Get a list of areas known to AI.
$CreatureAI.GetGameObject(ObjectId) => Get AI object by ID.
$CreatureAI.PartyManager => Get party manager.
$CreatureAI.ObjectManager => Get object manager.
$CreatureAI.Module => Get AI module object.

Note that calls cannot be made to the script interface when using
$CreatureAI to examine or update AI objects, that is, it is not
legal to use properties or methods that would cause the
ACR_CreatureBehavior system to try and make a call through the
script interface.

Only properties and data locally maintained by the creature AI
system can, thus, be examined or edited.