Resynchronize GUI state to new server in server to server portal.
When a player transfers servers with a portal, the game client
keeps its old GUI state (for the most part), but the server is not
informed of this.  This can cause the client-side and server-side
parts that manage the client GUI to get out of sync.

To counteract this problem, the portal initiating server now sends
a request to the portal destination server, informing it of how to
update its server-side GUI state to match the GUI state of the
player on the portal initiating server.

This fixes a problem where, on portal, the player has the wrong
information displayed in the chat select GUI, if it was open at
portal time.  The framework can be extended to work with other GUI
elements later, if that should be necessary, by creating additional
ResyncFlags bits (or additional ResyncState tokens).

The sequence for the resynchronization process is as so :

1) Source server sends an ACR_SERVER_IPC_EVENT_RUN_SCRIPT request
   to the destination server.  The script to run is
	"acr_resync_gui", and the script argument is a serialized
	ACR_ServerCommunicator.GUIResynchronizer.ResyncState.  This
	process is initiated by a call to
	ACR_ServerCommunicator.GUIResynchronizer.SendGUIStateToServer
	just before the ActivatePortal request is sent to the client
	attempting to portal.

2) Destination server receives the run script event and executes
   the script.  acr_resync_gui calls the server communicator with
	request code ACR_SERVER_IPC_HANDLE_GUI_RESYNC.  This causes
	the server communicator to execute
	ACR_ServerCommunicator.GUIResynchronizer.HandleGUIResync.

Simultaneously, the client attempts to portal to the destination
server.  The client may arrive before or after step 2.  If before,
the client has the GUI state applied when HandleGUIResync is first
called.  If after, the client has GUI state applied when the
server communicator detects the ClientEnter event for the incoming
player and subsequently invokes the
ACR_ServerCommunicator.GUIResynchronizer.OnClientEnter handler.

In turn, once the player has finished loading an area and is not
script hidden anymore, both roads lead to
ACR_ServerCommunicator.GUIResynchronizer.ResynchronizePlayerState,
which updates the PlayerState of the player to match the source
server's state for the appropriate GUI elements.  The client-side
GUI is then sent updates, as appropriate.