Make shared assembly resources work in all build configurations.
Shared assemblies that will be packaged as a binary resource
within ACR_AssemblyLoader now have an extra level of indirection;
the project output files are copied to release\SharedAssemblies
which is then referenced by the .resx files.

This ensures that per-build-configuration paths are abstracted
away from the .resx file, which does not understand project-level
dependencies for files generated as a project output.