Add further comment about FIleUpdateMethodDirectReplace.
Files updated with this method are expected to be not critical
to server startup.  Their contents are typically wholly managed
by the patch system (e.g. supplemental override scripts), and if
a rollback occurs the OLD file will be deleted entirely, assumed
to be laid back down again by the next patch attempt.

This implies that the server must be able to start up successfully
if a direct replaced file is not present.  So, this method cannot
be used for critical data files like haks, tlks, etc.  It is
primarily intended for the supplemental override script scenario,
used to insert a new script (like the "acr_patch_initialize"
script run by the server communicator at server startup) used to
adjust module behavior without requiring a whole HAK re-release.