Automatically correct game difficulty to canonical value.
This is driven off of a new config table variable, called
GameDifficultyLevel.  By default this is set to 3, that is, the
nwscript.nss manifest constant GAME_DIFFICULTY_CORE_RULES.

If the difficulty level is adjusted, then one of the two counters
GAME_DIFFICULTY_INCREMENT or GAME_DIFFICULTY_DECREMENT is updated
in the database, so as to allow tracking of whether servers have
a difficulty level configuration problem on a regular basis.

If a DM needs to manually override difficulty on a temporary basis,
then it is necessary to use the following scriptlet to set the
target difficulty as well, assuming 2 is the target difficulty,
via the "#sa rs <scriptlet> facility:"

$s.SetLocalInt($s.GetModule(), "ACR_GAME_DIFFICULTY_LEVEL", 2);
$s.SetGameDifficultyLevel(2);