Add #recorddata <character name|player name> command.
This command shows account statistics for a player, or a player and
character (the latter only if a character name is supplied).  Only
"public" information is provided.  The command is useful to quickly
identify if a player is a new player or is a member from in the
game for support purposes.