Document acr_set_player_password script further.
This script could also be manually invoked via a direct database
query to remotely reconfigure a server, if needed in a one-off
case (avoiding the need for a server restart).  Note that doing so
does not change the configuration file default for the server that
would be assigned on next reboot, in the absence of another
acr_set_player_password run remote script request.