Add module offline mode and lock to computer name for servers.
Optionally, a server can be locked to a particular computer name
so that in recovery scenarios, if an automatically starting server
instance comes online unexpectedly, it does not conflict with an
already running backup server instance.

By default there is no computer name restriction.  To establish
one, one can update the servers table and set the MachineName
column to a computer name.  Only that computer can host the module
until the MachineName field is cleared.

An offlined server sends a message to IRC notifying the tech team
of this fact and prominently records this in the log file.  If the
MachineName restriction is cleared, then the offlined server will
notify the tech team via IRC and restart to allow the server to
start up normally.

While offlined, players cannot log on to the server because the
Azure vault stays offline (it is never initialized causing the
server vault plugin to hold logon messages indefinitely).