GameWorldManager: Do not process IPC events while module offlined.
If a module is administratively offlined due to starting on the
wrong computer, do not process IPC events as these must be picked
up by the actual server instance.

Note that the server will continue to poll for the computer name
restriction being cleared without having to process the IPC event
queue.